6. Complex types - mixed list1

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -6,7 +6,7 @@ from typing import List, Union
 
 
 # def sum_mixed_list(mix_lst: List[float | int]) -> float:
-def sum_mixed_list(mix_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mix_lst: List[Union[int, float]]) -> float:
 
     """
     takes a list mxd_lst of integers and floats and
